bytt fra constate til react context api for søknad context
