+++ src/frontend/sider/Fagsak/Behandling/Sider/RegistrerSøknad/RegistrerSøknad.tsx
@@ -9,7 +9,7 @@ import Annet from './Annet';
 import Barna from './Barna';
 import SøknadType from './SøknadType';
 import { useBehandling } from '../../../../../context/behandlingContext/BehandlingContext';
-import { useSøknad } from '../../../../../context/SøknadContext';
+import { useSøknadContext } from '../../../../../context/SøknadContext';
 import MålformVelger from '../../../../../komponenter/MålformVelger';
 import { BehandlingSteg } from '../../../../../typer/behandling';
 import Skjemasteg from '../Skjemasteg';
@@ -33,7 +33,7 @@ const RegistrerSøknad: React.FC = () => {
         skjema,
         søknadErLastetFraBackend,
         visBekreftModal,
-    } = useSøknad();
+    } = useSøknadContext();
 
     return (
         <StyledSkjemasteg
